requirements.txt
fastapi
uvicorn[stardard]
azure-data-tables